客户端整合spring
