+++ rpc-framework-spring/README.md
@@ -1,0 +1,5 @@
+# 整合spring
+基于spring 5.2.6.RELEASE 
+
+- 通过一个注解就可以实现远程接口的调用 
+  @RpcServiceScan("github.javaguide.spring.service")
