+++ rpc-framework-spring/pom.xml
@@ -13,10 +13,12 @@
 
 
     <dependencies>
+
         <dependency>
             <groupId>org.springframework</groupId>
             <artifactId>spring-context</artifactId>
             <version>5.2.6.RELEASE</version>
+            <scope>provided</scope>
         </dependency>
 
         <dependency>
