+++ rpc-framework-spring/src/main/java/github/javaguide/ClientProxy.java
@@ -5,7 +5,7 @@ import github.javaguide.remoting.transport.ClientTransport;
 import github.javaguide.remoting.transport.netty.client.NettyClientTransport;
 
 /**
- * @description:
+ * @description: 简单写一个方法获取代理对象，其实应该在simple框架里面提供一个接口获取代理对象
  * @author:lvxuhong
  * @date:2020/6/18
  */
